avery thing works auth isn't completed ! need js script for table from for table template

+++ commands/my_commands.py
@@ -2,6 +2,7 @@ from flask import Blueprint
 from models.users import Users
 from models.init_db import db
 from colorama import Fore
+import os
 
 my_commands_app = Blueprint('my_command', __name__)
 
@@ -9,7 +10,9 @@ my_commands_app = Blueprint('my_command', __name__)
 @my_commands_app.cli.command('init-db')
 def init_db():
     '''CLI command for init db'''
-    db.drop_all()
+    if os.path.exists('instance/db.db'):
+        db.drop_all()
+        print(Fore.YELLOW + 'Last database deleted' + Fore.RESET)
     db.create_all()
     print(Fore.GREEN + 'Db inited!!!' + Fore.RESET)
